+++ src/main/java/com/example/fooddelivery/restaurant/service/RestaurantService.java
@@ -2,6 +2,9 @@ package com.example.fooddelivery.restaurant.service;
 
 import com.example.fooddelivery.common.exception.DuplicateException;
 import com.example.fooddelivery.common.exception.NotFoundException;
+import com.example.fooddelivery.common.exception.UnauthorizedException;
+import com.example.fooddelivery.owner.domain.Owner;
+import com.example.fooddelivery.owner.repository.OwnerRepository;
 import com.example.fooddelivery.restaurant.domain.Restaurant;
 import com.example.fooddelivery.restaurant.dto.CreateRestaurantReqDto;
 import com.example.fooddelivery.restaurant.dto.RestaurantDetailResDto;
@@ -19,16 +22,20 @@ import java.util.stream.Collectors;
 public class RestaurantService {
 
     private final RestaurantRepository restaurantRepository;
+    private final OwnerRepository ownerRepository;
 
     @Autowired
-    public RestaurantService(RestaurantRepository restaurantRepository) {
+    public RestaurantService(RestaurantRepository restaurantRepository,
+        OwnerRepository ownerRepository) {
         this.restaurantRepository = restaurantRepository;
+        this.ownerRepository = ownerRepository;
     }
 
     @Transactional
-    public Long createRestaurant(CreateRestaurantReqDto reqDto) {
+    public Long createRestaurant(String identifier, CreateRestaurantReqDto reqDto) {
+        Owner owner = findOwnerByIdentifier(identifier);
         validateRestaurant(reqDto);
-        Restaurant restaurant = convertToRestaurant(reqDto);
+        Restaurant restaurant = convertToRestaurant(reqDto, owner);
         Restaurant saveRestaurant = restaurantRepository.save(restaurant);
         return saveRestaurant.getId();
     }
@@ -43,8 +50,14 @@ public class RestaurantService {
         return restaurantRepository.findByName(name).isPresent();
     }
 
-    private Restaurant convertToRestaurant(CreateRestaurantReqDto reqDto) {
-        return Restaurant.createRestaurant(reqDto.getName(), reqDto.getMinPrice(), reqDto.getDeliveryFee());
+    private Restaurant convertToRestaurant(CreateRestaurantReqDto reqDto, Owner owner) {
+        return new Restaurant(reqDto.getName(), reqDto.getMinPrice(), reqDto.getDeliveryFee(), owner);
+    }
+
+    private Owner findOwnerByIdentifier(String identifier) {
+        return ownerRepository.findByIdentifier(identifier).orElseThrow(
+            () -> new NotFoundException("사용자를 찾을 수 없습니다.")
+        );
     }
 
     @Transactional(readOnly = true)
@@ -76,9 +89,16 @@ public class RestaurantService {
     }
 
     @Transactional
-    public void updateRestaurant(UpdateRestaurantReqDto reqDto, Long restaurantId) {
+    public void updateRestaurant(String identifier, UpdateRestaurantReqDto reqDto, Long restaurantId) {
         Restaurant restaurant = findRestaurantById(restaurantId);
+        validateOwner(identifier, restaurant);
         updateRestaurant(restaurant, reqDto);
+    }
+
+    private void validateOwner(String identifier, Restaurant restaurant) {
+        if(!restaurant.isOwner(identifier)) {
+            throw new UnauthorizedException("식당 계정만 사용할 수 있는 기능입니다.");
+        }
     }
 
     private void updateRestaurant(Restaurant restaurant, UpdateRestaurantReqDto reqDto) {
@@ -86,7 +106,9 @@ public class RestaurantService {
     }
 
     @Transactional
-    public void deleteRestaurant(Long restaurantId) {
+    public void deleteRestaurant(String identifier, Long restaurantId) {
+        Restaurant restaurant = findRestaurantById(restaurantId);
+        validateOwner(identifier, restaurant);
         restaurantRepository.deleteById(restaurantId);
     }
 }
